dom/serviceworkers/test/test_post_message.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 982726 - Test service worker post message </title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script src="utils.js"></script>
<script class="testbody" type="text/javascript">
  var magic_value = "MAGIC_VALUE_123";
  var registration;
  function start() {
    return navigator.serviceWorker.register("message_posting_worker.js",
                                            { scope: "./sw_clients/" })
      .then(swr => waitForState(swr.installing, 'activated', swr))
      .then((swr) => registration = swr);
  }

  function unregister() {
    return registration.unregister().then(function(result) {
      ok(result, "Unregister should return true.");
    }, function(e) {
      dump("Unregistering the SW failed with " + e + "\n");
    });
  }


  function testPostMessage(swr) {
    var p = new Promise(function(res, rej) {
      window.onmessage = function(e) {
        if (e.data === "READY") {
          swr.active.postMessage(magic_value);
        } else if (e.data === magic_value) {
          ok(true, "Worker posted the correct value.");
          res();
        } else {
          ok(false, "Wrong value. Expected: " + magic_value +
                    ", got: " + e.data);
          res();
        }
      }
    });

    var content = document.getElementById("content");
    ok(content, "Parent exists.");

    iframe = document.createElement("iframe");
    iframe.setAttribute('src', "sw_clients/service_worker_controlled.html");
    content.appendChild(iframe);

    return p.then(() => content.removeChild(iframe));
  }

  function runTest() {
    start()
      .then(testPostMessage)
      .then(unregister)
      .catch(function(e) {
        ok(false, "Some test failed with error " + e);
      }).then(SimpleTest.finish);
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.exemptFromPerDomainMax", true],
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
  ]}, runTest);
</script>
</pre>
</body>
</html>
